+++ test/__init__.py
@@ -3,6 +3,24 @@ This package contains the code for testing pyqode.python.
 
 To run the test suite, just run the following command::
 
-    python3 runtests.py
+    tox
+
+To select a specific environment, use the -e option. E.g. to run coverage
+tests::
+
+    tox -e cov
+
+Here is the list of available test environments:
+
+    - py27-pyqt4
+    - py27-pyqt5
+    - py32-pyqt4
+    - py32-pyqt5
+    - py33-pyqt4
+    - py33-pyqt5
+    - py34-pyqt4
+    - py34-pyqt5
+    - cov
+    - pep8
 
 """